pyproject.toml
[project]
name = "legal-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography==43.0.0",
    "langchain-chroma==0.1.2",
    "langchain-community==0.2.11",
    "langchain-core==0.2.29",
    "langchain-openai==0.1.21",
    "langchain-text-splitters==0.2.2",
    "langgraph==0.2.16",
    "pdfplumber==0.11.4",
    "pydantic>=2.0.0",
    "pypdf==4.3.1",
    "python-dotenv==1.0.1",
    "pyyaml>=6.0.0",
    "rank-bm25==0.2.2",
    "sudachidict-core>=20241021",
    "sudachipy==0.6.8",
]
